# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get list of articles",
                "operationId": "7e370449b00151b6422003b99ee09b88",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Create an article",
                "operationId": "10c43ee1198c7892a0897ed8f28c1775",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "body": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article created successfully"
                    }
                }
            }
        },
        "/api/preference-articles/{keyword}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get articles by user preference and keyword",
                "operationId": "2dbc9b79c4fe53ebb3dc4d817d018fde",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Get list of authors",
                "operationId": "7221682a8b1f36e80fd12fd170950ca1",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "17d569550a6f526e6d36d1af313a1f61",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "c_password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "c_password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User register successfully."
                    },
                    "422": {
                        "description": "Validation Error."
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User login",
                "operationId": "13789c6e8de56f03890c8e718282efb5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User login successfully."
                    },
                    "401": {
                        "description": "Unauthorised"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Articles",
            "description": "Articles"
        },
        {
            "name": "Authors",
            "description": "Authors"
        },
        {
            "name": "Auth",
            "description": "Auth"
        }
    ]
}